cleanups part1

+++ src/services/produce/cultivar/mod.rs
@@ -8,4 +8,4 @@ pub mod models;
 mod utils;
 
 /// Number of images allowed to be uploaded per cultivar
-const CULTIVAR_MAX_IMAGE: usize = 1;
+const CULTIVAR_MAX_IMAGE: u8 = 1;
